test/safe_app: refactor adata tests into a single file
Instead of manintaing separate files with the same tests for the
append-only data variants, refactor the tests in such a way that the
same code is looped against the different variants using conditional
staments to branch out whereever required
