+++ safe_authenticator/Cargo.toml
@@ -25,7 +25,7 @@ new_rand = { package = "rand", version = "0.6" }
 routing = { package = "mock_routing", path = "../mock_routing" }
 rust_sodium = "~0.10.2"
 safe_core = { path = "../safe_core", version = "~0.32.1" }
-safe-nd = "~0.2.0"
+safe-nd = { git = "https://github.com/maidsafe/safe-nd" }
 serde = "~1.0.27"
 serde_derive = "~1.0.27"
 threshold_crypto = "~0.3.2"
